renthouseapi/requirements.txt
alembic==1.13.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.37.0
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.7.1
Django==4.2.7
django-ckeditor==6.7.0
django-filter==24.2
django-js-asset==2.1.0
django-oauth-toolkit==2.4.0
django-sslserver==0.22
djangorestframework==3.15.1
drf-yasg==1.21.7
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
inflection==0.5.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.6
Mako==1.3.0
Markdown==3.6
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==24.1
passlib==1.7.4
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
python-slugify==8.0.1
python-version==0.0.2
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
text-unidecode==1.3
twilio==7.17.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0
alembic==1.13.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.37.0
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.7.1
Django==4.2.7
django-ckeditor==6.7.0
django-filter==24.2
django-js-asset==2.1.0
django-oauth-toolkit==2.4.0
django-sslserver==0.22
djangorestframework==3.15.1
drf-yasg==1.21.7
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
inflection==0.5.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.6
Mako==1.3.0
Markdown==3.6
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==24.1
passlib==1.7.4
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
python-slugify==8.0.1
python-version==0.0.2
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
text-unidecode==1.3
twilio==7.17.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0
